+++ src/DotNext.Tests/Net/Cluster/Consensus/Raft/NetworkTransport/ConnectionOriented/Tcp/TcpTransportTests.NoAot.cs
@@ -1,0 +1,11 @@
+using Microsoft.Extensions.Logging;
+
+namespace DotNext.Net.Cluster.Consensus.Raft.NetworkTransport.ConnectionOriented.Tcp;
+
+using DotNext.Extensions.Logging;
+
+partial class TcpTransportTests
+{
+    private static ILoggerFactory CreateDebugLoggerFactory(int port)
+        => TestLoggers.CreateDebugLoggerFactory(port.ToString(), static builder => builder.SetMinimumLevel(LogLevel.Debug));
+}