+++ src/DotNext.Tests/Net/Cluster/Consensus/Raft/TestLogEntry.cs
@@ -8,8 +8,6 @@ using TextMessage = Messaging.TextMessage;
 [ExcludeFromCodeCoverage]
 internal sealed class TestLogEntry(string command) : TextMessage(command, "Entry"), IInputLogEntry
 {
-    public DateTimeOffset Timestamp { get; } = DateTimeOffset.UtcNow;
-
     public long Term { get; init; }
 
     bool ILogEntry.IsSnapshot => false;