+++ src/cluster/DotNext.Net.Cluster/Net/Cluster/Consensus/Raft/StateMachine/WriteAheadLog.Flusher.cs
@@ -14,58 +14,87 @@ partial class WriteAheadLog
     private readonly AsyncAutoResetEventSlim? flushTrigger, flushCompleted;
     private readonly Task flusherTask;
     private readonly WeakReference<Task?> cleanupTask = new(target: null, trackResurrection: false);
+    private readonly bool flushOnCommit;
     
     private Checkpoint checkpoint;
-    private long commitIndex; // Commit lock protects modification of this field
-    private long flusherPreviousIndex, flusherOldSnapshot;
+    private Atomic<FlushState> flusherState;
+    private long flusherOldSnapshot;
 
-    private async Task FlushAsync<T>(T flushTrigger, CancellationToken token)
+    private async Task FlushAsync<T>(T trigger, CancellationToken token)
         where T : struct, IFlushTrigger
     {
         if (T.IsBackground)
+        {
             await Task.Yield();
+        }
         
-        // Weak ref tracks the task, but allows GC to collect associated state machine
-        // as soon as possible. While the task is running, it cannot be collected, because it's referenced
-        // by the async state machine.
         try
         {
-            if (T.IsBackground)
-            {
-                flusherOldSnapshot = SnapshotIndex;
-            }
-
             while (!token.IsCancellationRequested && backgroundTaskFailure is null)
             {
-                var newSnapshot = SnapshotIndex;
-                var newIndex = LastCommittedEntryIndex;
-
-                if (newIndex >= flusherPreviousIndex)
+                var newSnapshot = SnapshotIndex; // everything is flushed and committed up to this index
+                var lastCommittedIndex = long.Max(newSnapshot, LastCommittedEntryIndex);
+                var currentState = flusherState.Value;
+                var newIndex = long.Max(LastEntryIndex, newSnapshot);
+                var flushFrom = long.Max(newSnapshot + 1L, currentState.UnflushedIndex);
+                
+                if (newIndex >= flushFrom || lastCommittedIndex != currentState.FlushedCommitIndex)
                 {
                     // Ensure that the flusher is not running with the snapshot installation process concurrently
                     lockManager.SetCallerInformation("Flush Pages");
                     await lockManager.AcquireReadLockAsync(token).ConfigureAwait(false);
+                    
+                    // Can be modified by AppendAsync which supports rewriting of the tail,
+                    // that's why we need to read it again
+                    currentState = flusherState.Value;
+                    newIndex = long.Max(LastEntryIndex, newSnapshot);
+                    flushFrom = long.Max(newSnapshot + 1L, currentState.UnflushedIndex);
+                    var ts = default(Timestamp);
                     try
                     {
-                        var ts = new Timestamp();
-                        await Flush(flusherPreviousIndex, newIndex, token).ConfigureAwait(false);
+                        if (newIndex >= flushFrom)
+                        {
+                            ts = new();
+                            await FlushAsync(flushFrom, newIndex, token).ConfigureAwait(false);
 
-                        // everything up to toIndex is flushed, save the commit index
-                        await checkpoint.UpdateAsync<CheckpointVersion0>(new(newIndex), token).ConfigureAwait(false);
-                        FlushDurationMeter.Record(ts.ElapsedMilliseconds);
+                            // everything up to newIndex is flushed, save the commit index
+                            await checkpoint.UpdateAsync<CheckpointVersion1>(new(lastCommittedIndex, newIndex), token)
+                                .ConfigureAwait(false);
+
+                            currentState.UnflushedIndex = newIndex + 1L;
+                        }
+                        else if (lastCommittedIndex != currentState.FlushedCommitIndex)
+                        {
+                            ts = new();
+                            
+                            // update commit index only
+                            await checkpoint.UpdateAsync<CheckpointVersion1>(new(lastCommittedIndex, newIndex), token)
+                                .ConfigureAwait(false);
+
+                            currentState.UnflushedIndex = flushFrom;
+                        }
+                        else
+                        {
+                            goto exit;
+                        }
+
+                        currentState.FlushedCommitIndex = lastCommittedIndex;
+                        flusherState.Write(in currentState);
+                        exit: ;
                     }
                     finally
                     {
+                        if (!ts.IsEmpty)
+                            FlushDurationMeter.Record(ts.ElapsedMilliseconds);
+                        
                         lockManager.ReleaseReadLock();
                     }
                 }
 
-                if ((!cleanupTask.TryGetTarget(out var task) || task.IsCompletedSuccessfully) && flusherOldSnapshot < newSnapshot)
-                    cleanupTask.SetTarget(CleanUpAsync(newSnapshot, lifetimeToken));
+                StartCleanupIfNeeded(newSnapshot);
 
-                flushTrigger.NotifyCompleted();
-                Atomic.Write(ref flusherPreviousIndex, long.Max(flusherOldSnapshot = newSnapshot, newIndex) + 1L);
-                if (!await flushTrigger.WaitAsync(token).ConfigureAwait(false))
+                trigger.NotifyCompleted();
+                if (!await trigger.WaitAsync(token).ConfigureAwait(false))
                     break;
             }
         }
@@ -80,11 +109,23 @@ partial class WriteAheadLog
         }
         finally
         {
-            flushTrigger.Dispose();
+            trigger.Dispose();
         }
     }
 
-    private Task Flush(long fromIndex, long toIndex, CancellationToken token)
+    private void StartCleanupIfNeeded(long snapshotIndex)
+    {
+        // Weak ref tracks the task, but allows GC to collect associated state machine
+        // as soon as possible. While the task is running, it cannot be collected, because it's referenced
+        // by the async state machine.
+        if ((!cleanupTask.TryGetTarget(out var task) || task.IsCompletedSuccessfully) && flusherOldSnapshot < snapshotIndex)
+        {
+            cleanupTask.SetTarget(CleanUpAsync(snapshotIndex, lifetimeToken));
+            flusherOldSnapshot = snapshotIndex;
+        }
+    }
+
+    private Task FlushAsync(long fromIndex, long toIndex, CancellationToken token)
     {
         var metadataTask = metadataPages.FlushAsync(fromIndex, toIndex, token).AsTask();
 
@@ -96,7 +137,8 @@ partial class WriteAheadLog
         return Task.WhenAll(metadataTask, dataTask);
     }
 
-    private async Task EnsureFlushedAsync(CancellationToken token)
+    private async Task EnsureFlushedAsync<TChecker>(TChecker checker, CancellationToken token)
+        where TChecker : struct, IFlushStateChecker
     {
         Debug.Assert(flushCompleted is not null);
 
@@ -104,7 +146,7 @@ partial class WriteAheadLog
         var registration = linkedTokenSource.Token.UnsafeRegister(Signal, flushCompleted);
         try
         {
-            while (Atomic.Read(in flusherPreviousIndex) < LastCommittedEntryIndex)
+            while (checker.IsNotFlushed(flusherState.Read<long, TChecker>()))
             {
                 await flushCompleted.WaitAsync().ConfigureAwait(false);
                 if (linkedTokenSource.Token.IsCancellationRequested)
@@ -142,13 +184,15 @@ partial class WriteAheadLog
     public Task FlushAsync(CancellationToken token = default)
         => flushCompleted is null
             ? FlushAsync<ForegroundTrigger>(new(), token)
-            : EnsureFlushedAsync(token);
+            : flushOnCommit
+                ? EnsureFlushedAsync(new FlushedIndexChecker(LastEntryIndex), token)
+                : EnsureFlushedAsync(new FlushedCommitIndexChecker(LastCommittedEntryIndex), token);
 
     /// <inheritdoc cref="IAuditTrail.LastCommittedEntryIndex"/>
     public long LastCommittedEntryIndex
     {
-        get => Atomic.Read(in commitIndex);
-        private set => Atomic.Write(ref commitIndex, value);
+        get => Atomic.Read(in field);
+        private set => Atomic.Write(ref field, value);
     }
     
     private long Commit(long index)
@@ -173,6 +217,26 @@ partial class WriteAheadLog
         flushTrigger?.Set();
         CommitRateMeter.Add(count, measurementTags);
     }
+
+    [MethodImpl(MethodImplOptions.AggressiveInlining)]
+    private void RequestFlushIfNeeded()
+    {
+        if (flushOnCommit)
+        {
+            Debug.Assert(flushTrigger is not null);
+            
+            flushTrigger.Set();
+        }
+    }
+
+    private long UnflushedIndex
+    {
+        set
+        {
+            using var scope = flusherState.EnterLock();
+            scope.Value.UnflushedIndex = value;
+        }
+    }
     
     private interface IFlushTrigger : IDisposable
     {
@@ -184,20 +248,12 @@ partial class WriteAheadLog
     }
 
     [StructLayout(LayoutKind.Auto)]
-    private readonly struct BackgroundTrigger : IFlushTrigger
+    private readonly struct BackgroundTrigger(AsyncAutoResetEventSlim resetEvent, AsyncAutoResetEventSlim notification) : IFlushTrigger
     {
-        private readonly AsyncAutoResetEventSlim flushTrigger, flushNotification;
-        
-        public BackgroundTrigger(AsyncAutoResetEventSlim resetEvent, out AsyncAutoResetEventSlim notification)
-        {
-            flushTrigger = resetEvent;
-            notification = flushNotification = new();
-        }
-        
         ValueTask<bool> IFlushTrigger.WaitAsync(CancellationToken token)
-            => flushTrigger.WaitAsync();
+            => resetEvent.WaitAsync();
 
-        void IFlushTrigger.NotifyCompleted() => flushNotification.Set();
+        void IFlushTrigger.NotifyCompleted() => notification.Set();
 
         void IDisposable.Dispose()
         {
@@ -242,5 +298,36 @@ partial class WriteAheadLog
         {
             // nothing to do
         }
+    }
+    
+    [StructLayout(LayoutKind.Auto)]
+    private struct FlushState
+    {
+        public long UnflushedIndex, FlushedCommitIndex;
+    }
+    
+    private interface IFlushStateChecker : Atomic<FlushState>.IFieldReference<long>
+    {
+        bool IsNotFlushed(long index);
+    }
+
+    [StructLayout(LayoutKind.Auto)]
+    private readonly struct FlushedIndexChecker(long index) : IFlushStateChecker
+    {
+        bool IFlushStateChecker.IsNotFlushed(long unflushedIndex)
+            => unflushedIndex < index;
+
+        static ref readonly long Atomic<FlushState>.IFieldReference<long>.GetFieldReference(in FlushState value)
+            => ref value.UnflushedIndex;
+    }
+
+    [StructLayout(LayoutKind.Auto)]
+    private readonly struct FlushedCommitIndexChecker(long commitIndex) : IFlushStateChecker
+    {
+        bool IFlushStateChecker.IsNotFlushed(long flushedCommitIndex)
+            => flushedCommitIndex < commitIndex;
+
+        static ref readonly long Atomic<FlushState>.IFieldReference<long>.GetFieldReference(in FlushState value)
+            => ref value.FlushedCommitIndex;
     }
 }